requirements.txt
asgiref==3.9.2
Django==5.2.6
django-appconf==1.1.0
django-phonenumber-field==8.2.0
django-select2==8.4.3
phonenumbers==9.0.15
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.15.0
